requirement.txt
asgiref==3.3.4
Django==2.2.10
djangorestframework==3.10.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1